backend/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.5.64
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0

# Web framework
Flask>=2.0.0
Flask-Cors>=3.0.10
Werkzeug>=2.0.0
requests>=2.28.0

# Image processing and ML
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.21.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.65.0
typing-extensions>=4.5.0

# Note: Detectron2 should be installed separately using:
# python -m pip install 'git+https://github.com/facebookresearch/detectron2.git' 